refac: Move ob run wrapper into outerbounds module
