+++ src/outerbounds/mod.rs
@@ -1,5 +1,7 @@
 mod app;
 mod init;
+mod run;
 
 pub use app::{open_app, view_app};
-pub use init::{InitOptions, init_project, print_init_help};
+pub use init::{init_project, print_init_help, InitOptions};
+pub use run::run;
